requirements.txt
speechrecognition
PyAudio
pyttsx3
playsound
gTTS
yfinance
python-dotenv==1.0.0
